refactor(api): Self-managed database migrations
This is part of a larger effort to eventually migrate entirely from
go-pg to uptrace/bun. Bun has its own migration management but switching
from go-pg to bun invisibly is tricky. By rolling my own migration
system it will make the transition to bun easier since the table
tracking the migrations and the format of the migrations will not need
to change. This will also make it easier to eventually add support for a
separate track of SQLite migrations.
