+++ go.mod
@@ -20,7 +20,6 @@ require (
 	github.com/fsnotify/fsnotify v1.10.1
 	github.com/gavv/httpexpect/v2 v2.17.0
 	github.com/getsentry/sentry-go v0.46.2
-	github.com/go-pg/migrations/v8 v8.1.0
 	github.com/go-pg/pg/v10 v10.15.1
 	github.com/golang-jwt/jwt/v5 v5.3.1
 	github.com/gomodule/redigo v1.9.3
@@ -142,9 +141,6 @@ require (
 	mellium.im/sasl v0.3.1 // indirect
 	moul.io/http2curl/v2 v2.3.0 // indirect
 )
-
-// Fix migrations/v8 dependency
-replace github.com/go-pg/pg/v10 v10.4.0 => github.com/go-pg/pg/v10 v10.15.1
 
 // Keep everything on the same msgpack version
 replace github.com/vmihailenco/msgpack/v5 v5.3.4 => github.com/vmihailenco/msgpack/v5 v5.4.1
